Out of Memory fix (25-10-2016_12.42)

+++ app/src/main/java/com/group/chat/GroupChatActivity.java
@@ -5079,7 +5079,7 @@ public class GroupChatActivity extends FragmentActivity implements OnClickListen
                     break;
                 case R.id.c_audio_call:
                     if(!CallDispatcher.myStatus.equalsIgnoreCase("0")) {
-                        if (audio_call.getTag() == 1) {
+                        if ((int)audio_call.getTag() == 1) {
                             if(isPrivateBack){
                                 LL_privateReply.setVisibility(View.GONE);
                             }else {
